+++ tags.lua
@@ -89,22 +89,30 @@ local function get_texture(info, colorFull, colorEmpty)
     return barFull .. barEmpty
 end
 
+local function combine(info, value)
+    if value and value ~= "" then
+        return (info.prefix or "") .. value .. (info.suffix or "")
+    else
+        return ""
+    end
+end
+
 Addon.TAGS.Definition = {
     ["name"] = {
         desc = "Name of the faction",
-        value = function(info) return Addon.CONST.MESSAGE_COLORS.NAME .. info.name .. "|r" end
+        value = function(info) return Addon.CONST.MESSAGE_COLORS.NAME .. combine(info, info.name) .. "|r" end
     },
     ["c_name"] = {
         desc = "Name of the faction colored by standing",
-        value = function(info) return info.standingColor .. info.name .. "|r" end
+        value = function(info) return info.standingColor .. combine(info, info.name) .. "|r" end
     },
     ["standing"] = {
         desc = "Current reputation standing",
         value = function(info)
             if Addon.db.profile.Reputation.showParagonCount and info.paragon ~= "" then
-                return info.standingText .. " (" .. info.paragon .. ")"
+                return combine(info, info.standingText .. " (" .. info.paragon .. ")")
             else
-                return info.standingText
+                return combine(info, info.standingText)
             end
         end
     },
@@ -113,9 +121,9 @@ Addon.TAGS.Definition = {
         value = function(info)
             local standingTextShort = first_letters(info.standingText, Addon.db.profile.Reputation.shortCharCount)
             if Addon.db.profile.Reputation.showParagonCount and info.paragon ~= "" then
-                return standingTextShort .. " (" .. info.paragon .. ")"
+                return combine(info, standingTextShort .. " (" .. info.paragon .. ")")
             else
-                return standingTextShort
+                return combine(info, standingTextShort)
             end
         end
     },
@@ -125,9 +133,9 @@ Addon.TAGS.Definition = {
             if Addon.db.profile.Reputation.showParagonCount and info.paragon ~= "" then
                 local reputationColors = Addon.db.profile.Colors
                 local paragonColor = ("|cff%.2x%.2x%.2x"):format(reputationColors[9].r*255, reputationColors[9].g*255, reputationColors[9].b*255)
-                return info.standingColor .. info.standingText .. "|r" .. paragonColor .. " (" .. info.paragon .. ")|r"
+                return info.standingColor .. combine(info, info.standingText) .. "|r" .. paragonColor .. " (" .. info.paragon .. ")|r"
             else
-                return info.standingColor .. info.standingText .. "|r"
+                return info.standingColor .. combine(info, info.standingText) .. "|r"
             end
         end
     },
@@ -138,82 +146,82 @@ Addon.TAGS.Definition = {
             if Addon.db.profile.Reputation.showParagonCount and info.paragon ~= "" then
                 local reputationColors = Addon.db.profile.Colors
                 local paragonColor = ("|cff%.2x%.2x%.2x"):format(reputationColors[9].r*255, reputationColors[9].g*255, reputationColors[9].b*255)
-                return info.standingColor .. standingTextShort .. "|r" .. paragonColor .. " (" .. info.paragon .. ")|r"
+                return info.standingColor .. combine(info, standingTextShort) .. "|r" .. paragonColor .. " (" .. info.paragon .. ")|r"
             else
-                return info.standingColor .. standingTextShort .. "|r"
+                return info.standingColor .. combine(info, standingTextShort) .. "|r"
             end
         end
     },
     ["change"] = {
         desc = "Actual gain/loss of reputation",
-        value = function(info) return (info.negative and "-" or "+") .. info.change end
+        value = function(info) return combine(info, (info.negative and "-" or "+") .. info.change) end
     },
     ["c_change"] = {
         desc = "Actual gain/loss of reputation (green for gain, red for loss)",
-        value = function(info) return (info.negative and Addon.CONST.MESSAGE_COLORS.NEGATIVE or Addon.CONST.MESSAGE_COLORS.POSITIVE) .. (info.negative and "-" or "+") .. info.change .. "|r" end
+        value = function(info) return (info.negative and Addon.CONST.MESSAGE_COLORS.NEGATIVE or Addon.CONST.MESSAGE_COLORS.POSITIVE) .. combine(info, (info.negative and "-" or "+") .. info.change ).. "|r" end
     },
     ["session"] = {
         desc = "Gain of reputation in current session",
-        value = function(info) return ((info.session > 0) and "+" or "") .. info.session end
+        value = function(info) return combine(info, ((info.session > 0) and "+" or "") .. info.session) end
     },
     ["c_session"] = {
         desc = "Gain of reputation in current session (green for gain, red for loss)",
-        value = function(info) return ((info.session > 0) and Addon.CONST.MESSAGE_COLORS.POSITIVE or Addon.CONST.MESSAGE_COLORS.NEGATIVE) .. ((info.session > 0) and "+" or "") .. info.session .. "|r" end
+        value = function(info) return ((info.session > 0) and Addon.CONST.MESSAGE_COLORS.POSITIVE or Addon.CONST.MESSAGE_COLORS.NEGATIVE) .. combine(info, ((info.session > 0) and "+" or "") .. info.session) .. "|r" end
     },
     ["current"] = {
         desc = "Current reputation value",
-        value = function(info) return info.current end
+        value = function(info) return combine(info, info.current) end
     },
     ["next"] = {
         desc = "Reputation boundary for next level",
-        value = function(info) return info.maximum end
+        value = function(info) return combine(info, info.maximum) end
     },
     ["bottom"] = {
         desc = "Minimum reputation in current standing",
-        value = function(info) return info.bottom end
+        value = function(info) return combine(info, info.bottom) end
     },
     ["top"] = {
         desc = "Maximum reputation in current standing",
-        value = function(info) return info.top end
+        value = function(info) return combine(info, info.top) end
     },
     ["toGo"] = {
         desc = "Reputation to gain/loss for next/previous standing",
-        value = function(info) return (info.negative and ("-" .. info.current) or (info.maximum - info.current)) end
+        value = function(info) return combine(info, (info.negative and ("-" .. info.current) or (info.maximum - info.current))) end
     },
     ["changePercent"] = {
         desc = "Percentual change of reputation",
-        value = function(info) return format("%.2f%%%%", (info.maximum == 0 and 0) or (info.change / info.maximum * 100)) end
+        value = function(info) return combine(info, format("%.2f%%", (info.maximum == 0 and 0) or (info.change / info.maximum * 100))) end
     },
     ["sessionPercent"] = {
         desc = "Percentual change of reputation during active session",
-        value = function(info) return format("%.1f%%%%", (info.maximum == 0 and 0) or (info.session / info.maximum * 100)) end
+        value = function(info) return combine(info, format("%.1f%%", (info.maximum == 0 and 0) or (info.session / info.maximum * 100))) end
     },
     ["currentPercent"] = {
         desc = "Percent of next standing",
-        value = function(info) return format("%.1f%%%%", (info.maximum == 0 and 0) or (info.current / info.maximum * 100)) end
+        value = function(info) return combine(info, format("%.1f%%", (info.maximum == 0 and 0) or (info.current / info.maximum * 100))) end
     },
     ["paragonLevel"] = {
         desc = "Paragon level (with reward icon if available)",
-        value = function(info) return info.paragon end
+        value = function(info) return combine(info, info.paragon) end
     },
     ["c_paragonLevel"] = {
         desc = "Colored paragon level (with reward icon if available)",
         value = function(info)
             local reputationColors = Addon.db.profile.Colors
             local paragonColor = ("|cff%.2x%.2x%.2x"):format(reputationColors[9].r*255, reputationColors[9].g*255, reputationColors[9].b*255)
-            return paragonColor .. info.paragon .. "|r"
+            return paragonColor .. combine(info, info.paragon) .. "|r"
         end
     },
     ["renownLevel"] = {
         desc = "Renown level",
-        value = function(info) return info.renown end
+        value = function(info) return combine(info, info.renown) end
     },
     ["c_renownLevel"] = {
         desc = "Colored renown level",
         value = function(info)
             local reputationColors = Addon.db.profile.Colors
             local renownColor = ("|cff%.2x%.2x%.2x"):format(reputationColors[10].r*255, reputationColors[10].g*255, reputationColors[10].b*255)
-            return renownColor .. info.renown .. "|r"
+            return renownColor .. combine(info, info.renown) .. "|r"
         end
     },
     ["bar"] = {
@@ -245,9 +253,9 @@ Addon.TAGS.Definition = {
         value = function(info)
             if info.change then
                 if info.negative then
-                    return math.ceil(info.current/info.change)
+                    return combine(info, math.ceil(info.current/info.change))
                 else
-                    return math.ceil((info.maximum - info.current)/info.change)
+                    return combine(info, math.ceil((info.maximum - info.current)/info.change))
                 end
             end
             return ""
@@ -281,13 +289,13 @@ Addon.TAGS.Definition = {
     ["signText"] = {
         desc = "If a reputation is gained then it will display 'increased' when lost then 'decreased'",
         value = function(info)
-            return (info.negative and "decreased") or "increased"
+            return combine(info, (info.negative and "decreased") or "increased")
         end
     },
     ["standingNext"] = {
         desc = "Shows next/previous (depending on gain or loss) standing/paragon",
         value = function(info)
-            return info.standingTextNext
+            return combine(info, info.standingTextNext)
         end
     },
     ["c_standingNext"] = {
@@ -296,16 +304,16 @@ Addon.TAGS.Definition = {
             if info.standingIdNext then
                 local reputationColors = Addon.db.profile.Colors
                 local standingColor = ("|cff%.2x%.2x%.2x"):format(reputationColors[info.standingIdNext].r*255, reputationColors[info.standingIdNext].g*255, reputationColors[info.standingIdNext].b*255)
-                return standingColor .. info.standingTextNext .. "|r"
+                return standingColor .. combine(info, info.standingTextNext) .. "|r"
             else
-                return info.standingColor
+                return combine(info, info.standingtextNext)
             end
         end
     },
     ["standingNextShort"] = {
         desc = "Shows next/previous (depending on gain or loss) standing/paragon, with a maximum of 'x' characters per word, can be set in the options (1 is the default value for x).",
         value = function(info)
-            return first_letters(info.standingTextNext, Addon.db.profile.Reputation.shortCharCount)
+            return combine(info, first_letters(info.standingTextNext, Addon.db.profile.Reputation.shortCharCount))
         end
     },
     ["c_standingNextShort"] = {
@@ -314,9 +322,9 @@ Addon.TAGS.Definition = {
             if info.standingIdNext then
                 local reputationColors = Addon.db.profile.Colors
                 local standingColor = ("|cff%.2x%.2x%.2x"):format(reputationColors[info.standingIdNext].r*255, reputationColors[info.standingIdNext].g*255, reputationColors[info.standingIdNext].b*255)
-                return standingColor .. first_letters(info.standingTextNext, Addon.db.profile.Reputation.shortCharCount) .. "|r"
+                return standingColor .. combine(info, first_letters(info.standingTextNext, Addon.db.profile.Reputation.shortCharCount)) .. "|r"
             else
-                return first_letters(info.standingTextNext, Addon.db.profile.Reputation.shortCharCount)
+                return combine(info, first_letters(info.standingTextNext, Addon.db.profile.Reputation.shortCharCount))
             end
         end
     },