+++ CHANGELOG.md
@@ -1,6 +1,24 @@
- # v1.2.3-beta1 ()
+ # v1.3.0-beta1 ()
 
-## Update
+## **Conditional prefixes and suffixes for TAGS**
+I have added functionality that allows adding conditional prefixes and/or suffixes to any text TAG (except for bar, icon, etc.).
+
+The format is as follows:
+
+```
+[{prefix}TAG{suffix}]
+```
+There cannot be any spaces between the opening bracket, {prefix}, TAG, {suffix}, and closing bracket. However, prefix and suffix can be any text.
+
+**Also, if `[TAG]` evaluates to an empty value, neither the prefix nor the suffix will be displayed.**
+
+> Here are a few examples:
+
+- `[{Renown level: }renownLevel]` = "Renown level: 25" (If the faction isn't renowned, nothing will be displayed!)
+- `[{Next rank: }standingNext]` = "Next rank: Honored" (Again, if [standingNext] does not return a value, nothing will be displayed.)
+- `[{Level }paragonLevel{ paragon}]` = "Level 5 paragon" (Nothing will be displayed if there is no Paragon level available.)
+
+## Added
 - Added TAGs `[standingColorStart]` and `[standingColorEnd]`. They will color the text between them in the message with a standing color. **These have to be used in pairs** and `[standingColorStart]` must precede `[standingColorEnd]`. (requested by filliph)
 - Added a TAG `[signText]`. It will display "increased" when a reputation is gained and "decreased" when it is lost.
 - Added group of TAGs `[standingNext]` (with `[c_...]` and `[...Short]` variants). They will display next standing/renown/paragon standing. They only work for standard  *... -> "Neutral" -> "Friendly" -> "Honored" -> ...* progression or for renown and paragon factions.